Fixed bug, changed name of mod...
Fixed bug that would prevent using a SEED1 value greator then 1000.
Didn't set arg[0] value correctly. Mistanly used arg[1] for the check
which is used for SEED2.

Also made slight change to name of mod to better match the name of the
game. the main command is now "gungeonseed" instead of "dungeonseed".